+++ modules/ml-training/k8s-managed/app.py
@@ -1,18 +1,28 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 import os
+from typing import cast
 
 from aws_cdk import App, CfnOutput, Environment
 
 from stack import TrainingPipeline
 
+# Project specific
 deployment_name = os.environ["ADDF_DEPLOYMENT_NAME"]
 module_name = os.environ["ADDF_MODULE_NAME"]
-eks_cluster_name = os.environ["ADDF_PARAMETER_EKS_CLUSTER_NAME"]
-eks_admin_role_arn = os.environ["ADDF_PARAMETER_EKS_CLUSTER_ADMIN_ROLE_ARN"]
-eks_oidc_provider_arn = os.environ["ADDF_PARAMETER_EKS_OIDC_ARN"]
-training_namespace_name = os.environ["ADDF_PARAMETER_TRAINING_NAMESPACE_NAME"]
-training_image_uri = os.environ["ADDF_PARAMETER_TRAINING_IMAGE_URI"]
+
+
+def _param(name: str) -> str:
+    return f"ADDF_PARAMETER_{name}"
+
+
+eks_cluster_name = os.getenv(_param("EKS_CLUSTER_NAME"))
+eks_admin_role_arn = os.getenv(_param("EKS_CLUSTER_ADMIN_ROLE_ARN"))
+eks_oidc_provider_arn = os.getenv(_param("EKS_OIDC_ARN"))
+eks_cluster_endpoint = os.getenv(_param("EKS_CLUSTER_ENDPOINT"))
+eks_cert_auth_data = os.getenv(_param("EKS_CERT_AUTH_DATA"))
+training_namespace_name = os.getenv(_param("TRAINING_NAMESPACE_NAME"))
+training_image_uri = os.getenv(_param("TRAINING_IMAGE_URI"))
 
 app = App()
 
@@ -21,11 +31,13 @@ stack = TrainingPipeline(
     id=f"addf-{deployment_name}-{module_name}",
     deployment_name=deployment_name,
     module_name=module_name,
-    eks_cluster_name=eks_cluster_name,
-    eks_admin_role_arn=eks_admin_role_arn,
-    eks_openid_connect_provider_arn=eks_oidc_provider_arn,
-    training_namespace_name=training_namespace_name,
-    training_image_uri=training_image_uri,
+    eks_cluster_name=cast(str, eks_cluster_name),
+    eks_admin_role_arn=cast(str, eks_admin_role_arn),
+    eks_openid_connect_provider_arn=cast(str, eks_oidc_provider_arn),
+    eks_cluster_endpoint=cast(str, eks_cluster_endpoint),
+    eks_cert_auth_data=cast(str, eks_cert_auth_data),
+    training_namespace_name=cast(str, training_namespace_name),
+    training_image_uri=cast(str, training_image_uri),
     env=Environment(
         account=os.environ["CDK_DEFAULT_ACCOUNT"],
         region=os.environ["CDK_DEFAULT_REGION"],